requirements.txt
# Core training stack (install as needed)
torch
torch-geometric
numpy
pyyaml
hydra-core>=1.3
omegaconf
